feat(ingest): refactor pipeline, add tests, and document Slack alerts
