+++ .github/workflows/ingest.yml
@@ -84,3 +84,38 @@ jobs:
           fi
           git commit -m "chore(ingest): refresh external documentation"
           git push
+
+      - name: Notify Slack
+        if: always() && secrets.SLACK_WEBHOOK_URL != ''
+        env:
+          SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
+          STATUS: ${{ job.status }}
+          RUN_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
+          SOURCE_ID: ${{ steps.register-self-service.outputs.source_id || github.event.inputs.source_id || github.event.inputs.source || github.event.client_payload.source_id || 'all-enabled-sources' }}
+          ACTOR: ${{ github.actor }}
+          REPOSITORY: ${{ github.repository }}
+          TRIGGER: ${{ github.event_name }}
+        shell: bash
+        run: |
+          set -euo pipefail
+
+          case "$STATUS" in
+            success)
+              ICON=":white_check_mark:"
+              ;;
+            failure)
+              ICON=":x:"
+              ;;
+            cancelled)
+              ICON=":warning:"
+              ;;
+            *)
+              ICON=":information_source:"
+              ;;
+          esac
+
+          TEXT="$ICON Ingestion *$STATUS* for *$REPOSITORY*\nSource: *$SOURCE_ID*\nTrigger: *$TRIGGER* by *$ACTOR*\nRun: $RUN_URL"
+
+          curl -sS -X POST "$SLACK_WEBHOOK_URL" \
+            -H "Content-Type: application/json" \
+            --data "{\"text\":\"${TEXT//$'\n'/\\n}\"}" >/dev/null
