+++ tests/unit/scripts/fixtures/repo/documentation/guide.md
@@ -1,0 +1,5 @@
+# Guide
+
+This page references the same asset again:
+
+![Diagram](images/diagram.png)
