+++ tests/unit/scripts/fixtures/repo/documentation/images/diagram.png
@@ -1,0 +1,1 @@
+fixture-png-bytes
