+++ tests/unit/scripts/fixtures/repo/documentation/index.html.md.erb
@@ -1,0 +1,9 @@
+---
+title: Home
+---
+
+# Welcome
+
+See the [guide](/documentation/guide.html).
+
+![Diagram](/documentation/images/diagram.png)
