+++ tests/unit/scripts/fixtures/repo/portal.yaml
@@ -1,0 +1,3 @@
+docs:
+  path: documentation
+owner_slack: "#portal-team"
